Workaround stream_select issues. Closes #26
Note that because of the `continue` on `line 91`, the input for the current line continues to be empty until at least 1 more enter and then a statement that returns a value. If `^C` is pressed several times in a row, the line will continue to build up to no known effect.